% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst_reversed.R
\name{jst_reversed}
\alias{jst_reversed}
\title{Run a Reversed Joint Sentiment Topic (JST) Model}
\usage{
jst_reversed(
  dfm,
  sentiLexInput = NULL,
  numSentiLabs = 3,
  numTopics = 10,
  numIters = 3,
  updateParaStep = -1,
  alpha = -1,
  beta = -1,
  gamma = -1,
  excludeNeutral = FALSE
)
}
\arguments{
\item{dfm}{A quanteda "dfm" (document-feature matrix) object containing the
text data.}

\item{sentiLexInput}{Optional: A quanteda dictionary object for
semi-supervised learning. If provided, the sentiment labels will be
inferred from this dictionary, overriding \code{numSentiLabs}. By default,
an additional neutral category is included. To exclude, set \code{excludeNeutral =
TRUE}.}

\item{numSentiLabs}{Number of sentiment labels. Default is 3.}

\item{numTopics}{Number of topics to extract from the data. Default is 10.}

\item{numIters}{Number of iterations for the Gibbs sampler. Set to 3 for test
runs. Adjust as needed for real data.}

\item{updateParaStep}{Number of iterations between updates of the
hyperparameter alpha.}

\item{alpha}{Hyperparameter for topic distributions. If not set, it will be
calculated using the formula provided.}

\item{beta}{Hyperparameter for sentiment-topic word distributions. Default is
based on sentiment dictionary presence.}

\item{gamma}{Hyperparameter for document-topic distributions. If not set, it
will be calculated using the formula provided.}

\item{excludeNeutral}{Boolean flag to determine if neutral sentiment category
should be excluded. Default is \code{FALSE}.}
}
\value{
A JST_reversed.result. See \link{JST_reversed.result}
}
\description{
This function implements the Reversed Joint Sentiment Topic model based on
the Gibbs sampling method. For the theoretical background, refer to the paper
by Lin et al. (2012).
}
\note{
For detailed model description, refer to: Lin, C., He, Y., Everson, R.
  & Ruger, S. (2012). Weakly supervised joint sentiment-topic detection from
  text. IEEE Transactions on Knowledge and Data engineering, 24(6),
  pp.1134-1145.
}
\examples{
\dontrun{
library(quanteda)
library(quanteda.textmodels)
data_irishbudget2010 <- data_corpus_irishbudget2010 \%>\%
  tokens() \%>\%
  dfm()

model <- jst_reversed(data_irishbudget2010, sentiLexInput = paradigm(), numTopics = 5, numIters = 150)
model
}

}
