% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstManyRuns.R
\name{jstManyRuns}
\alias{jstManyRuns}
\title{Estimate Sentiment Labels with the Joint Sentiment Topic Model Several Times}
\usage{
jstManyRuns(
  dfm,
  sentiLexInput = NULL,
  numSentiLabs = 3,
  numTopics = 10,
  numIters = 3,
  updateParaStep = -1,
  alpha = -1,
  beta = -1,
  gamma = -1,
  excludeNeutral = FALSE,
  n = 30,
  confidence = 0.95,
  ncores = NA,
  seed = NA
)
}
\arguments{
\item{dfm}{A quanteda "dfm" (document-feature matrix) object containing the
text data.}

\item{sentiLexInput}{Optional: A quanteda dictionary object for
semi-supervised learning. If provided, the sentiment labels will be
inferred from this dictionary, overriding \code{numSentiLabs}. By default,
an additional neutral category is included. To exclude, set \code{excludeNeutral =
TRUE}.}

\item{numSentiLabs}{Number of sentiment labels. Default is 3.}

\item{numTopics}{Number of topics to extract from the data. Default is 10.}

\item{numIters}{Number of iterations for the Gibbs sampler. Set to 3 for test
runs. Adjust as needed for real data.}

\item{updateParaStep}{Number of iterations between updates of the
hyperparameter alpha.}

\item{alpha}{Hyperparameter for topic distributions. If not set, it will be
calculated using the formula provided.}

\item{beta}{Hyperparameter for sentiment-topic word distributions. Default is
based on sentiment dictionary presence.}

\item{gamma}{Hyperparameter for document-topic distributions. If not set, it
will be calculated using the formula provided.}

\item{excludeNeutral}{Boolean flag to determine if neutral sentiment category
should be excluded. Default is \code{FALSE}.}

\item{n}{Number of model runs (defaults to 30)}

\item{confidence}{Confidence level for confidence intervals (defaults to 0.95)}

\item{ncores}{Number of (logical) CPU cores used for the estimation. (defaults to available number of cores - 3)}

\item{seed}{Seed for random number generator for reproducibility (defaults to \code{NA})}
}
\value{
A data.frame with mean sentiment label predictions and uncertainty estimates
}
\description{
This function runs the Joint Sentiment Topic (JST) model multiple times and
provides an average sentiment label prediction for each run.
}
\details{
Instead of running the JST model once, this function runs it 'n' times in
parallel. Instead of the full results, this function provides the average
sentiment label predictions with associated uncertainty estimates. For the
foundational model's details, please refer to the \code{\link{jst}}
function's documentation.

\strong{Advantages}:
\itemize{
  \item \strong{Improved Robustness}: Multiple runs can capture a more accurate representation of the underlying data structure by averaging out the noise.
  \item \strong{Uncertainty Estimation}: Provides a measure of variability in sentiment predictions, offering insights into the reliability of the results.
  \item \strong{Parallel Processing}: Utilizes parallel processing for faster computations, especially beneficial for large datasets or complex models.
}

\strong{Disadvantages}:
\itemize{
  \item \strong{Computational Cost}: Requires more computational resources due to multiple runs.
  \item \strong{Memory Usage}: Might increase memory demand especially if the dataset is large.
  \item \strong{Results Interpretation}: Provides average values rather than full model details, which might be limiting if detailed model analysis is needed.
}
}
\examples{
\dontrun{
library(quanteda)
library(quanteda.textmodels)
data_irishbudget2010 <- data_corpus_irishbudget2010 \%>\% 
  tokens() \%>\% 
  dfm()

model <- jstManyRuns(data_irishbudget2010, 
                     sentiLexInput = paradigm(), 
                     numTopics = 5, 
                     numIters = 150,
                     n = 5,
                     confidence = 0.95,
                     ncores = 4,
                     seed = 123456)
}

}
