% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentitopics.R
\docType{package}
\name{sentitopics-package}
\alias{sentitopics}
\alias{sentitopics-package}
\title{sentitopics: Joint Estimation of Sentiment and Topcis in Textual Data}
\description{
sentitopics performs Joint Sentiment Topic modelling and includes a
reversed model (for info on the latter, email me) and several
methods to evaluate the results and make them more accessible.
All models take \pkg{quanteda} dfm objects as inputs. If you prefer \pkg{tidytext},
tm objects are internally converted to dfm objects. Furthermore,
all models have a defined tidy method. See \pkg{broom} for more
information.
}
\author{
\strong{Maintainer}: Christian Pipal \email{c.pipal@uva.nl}

Authors:
\itemize{
  \item Max Boiten
}

Other contributors:
\itemize{
  \item Chenghua Lin [copyright holder]
  \item Yulan He [copyright holder]
}

}
